refactor: split view functions
