+++ view_importer.go
@@ -1,0 +1,10 @@
+package main
+
+// viewImporter renders the importer wizard view.
+func (m model) viewImporter() string {
+	m.ui.elemPos = map[string]int{}
+	if m.importWizard == nil {
+		return ""
+	}
+	return m.overlayHelp(m.importWizard.View())
+}
